+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.2] - DEV
-
+- Various optimizations.
+- Improve error messages.
 
 ## [0.1] - 2019-10-08
 ### Added